Better way of splitting ING "Mededelingen" field.
ING has a lot of different subfields in "Mededelingen" that broke my conversion. So I improved it by splitting up all the subfields (that I know of) and making a subselection of it and separate them with a nice pipeline (|) character.
If you want to customize the fields you want to keep simply edit line 262 (the third argument of memo: new Field()).